+++ crates/rtlola-frontend/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tuple Literal to MIR
 - Added `InstanceSelection::FilteredAll` and `InstanceSelection::FilteredFresh` variants for filtered instance aggregation.
 - Added `aggregates` field to mir
+- Added `origin` to windows
 
 ## [0.7.0] - 28.06.2024